UPSTREAM: ASoC: Intel: Skylake: Harden DSP detection with PCI class/subclass info
The existing PPCAP and GCAP fields cannot be used reliably to
determine if the DSP is enabled by the BIOS. Instead rely on the
class/subclass information to find out if this driver can run or
not. The values in the code don't seem to be documented in publicly
available documents but are part of recommendations made to BIOS
writers and have been verified to be accurate on a number of
platforms.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit c746de8dbc7b0ae9df491f7a99a6dab34203b51b)

BUG=b:110025789
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: I7f3ecac225c75b187f10238c8f5d4723f4c18f37
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>