FROMGIT: ASoC: audio-graph-card: tidyup prefix for snd_soc_codec_conf
Current audio-graph-card is handling "prefix" by many ways.
But, it is not useful and readable.
We want to do is that allow having it everywere.
This patch supports it.
It will be overwrote if lower node has it.

	sound {
		prefix = "xxx"; // initial
	};

	codec {
		audio-graph-card,prefix = "xxx"; // overwrite
		ports {
			prefix = "xxx"; // overwrite
			port {
				prefix = "xxx"; // overwrite
			};
		};
	};

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 66164a4da30dd295a1b60568c02bee37cfd9895a
 https://github.com/broonie/asoc for-next)

BUG=b:110025789
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: I4f4f0f361b29c8b2f947c0d62c28f6fc497248aa
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>