FROMGIT: ASoC: simple-card: tidyup mclk-fs method
Current simple-card is handling "mclk-fs" by many way.
But, it is not useful and readable.
We want to do is that allow having mclk-fs everywere.
This patch support it.
It will be overwrote if lower node has it.

sound {
	simple-audio-card,mclk-fs = <xxx>; // for initial

	simple-audio-card,dai-link {
		mclk-fs = <xxx>; // overwrite
		cpu {
			mclk-fs = <xxx>; // overwrite
		};
		codec {
			mclk-fs = <xxx>; // overwrite
		};
	};
};

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 79e834914d1fa2d078c586094ad347cc05aa428c
 https://github.com/broonie/asoc for-next)

BUG=b:110025789
TEST=Test Audio use cases.

Change-Id: Ief766881c0970d1c53ebb1fafcd7dd49d729d7d7
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>