FROMGIT: ASoC: simple-card: tidyup convert_rate/channel method
Current simple-card is handling "convert_rate/channel"
by many ways. But, it is not useful and readable.
We want to do is that allow having it everywere.
This patch support it.
It will be overwrote if lower node has it.

sound {
	simple-audio-card,convert_channels = <xxx>; // initial

	simple-audio-card,dai-link {
		convert_channels = <xxx>; // overwrite
		cpu {
			convert_channels = <xxx>; // overwrite
		};
		codec {
			convert_channels = <xxx>; // overwrite
		};
	};
};

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit dd82410f7e939f00591c415d62915a89ac2fb8f0
 https://github.com/broonie/asoc for-next)

BUG=b:110025789
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: Ibd0faf707e5c6e21c73a7ecfaac308275eb308e9
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>