FROMGIT: ASoC: es8316: Add jack-detect support
Adding jack-detect support may seem weird for a codec with only
a single output, but it is necessary. The ES8316 appnote showing
the intended usage uses a jack-receptacle which physically disconnects
the speakers from the output when a jack is plugged in.

But all 3 devices using the es8316 which I have (2 Cherry Trail
devices and one Bay Trail CR device), use an analog mux to disconnect
the speakers, driven by a GPIO. In order to enable/disable the speakers
at the right time, we need jack-detect.

The same goes for the microphone where we must correctly set the mux
for the single ADC to either the internal or the headset microphone.

All devices I have support the es8316's builtin jack-detect functionality.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 822257661031faa437336058d8a32bf1844ad9c6
 https://github.com/broonie/asoc for-next)

BUG=b:110025789
TEST=Test Audio use cases.

Change-Id: Iade6ea7875eef5222d1d5b2bc9ab045dee417c5c
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>