FROMGIT: ASoC: pcm512x: Implement the set_bclk_ratio interface
Some boards, such as the HiFiBerry DAC+ Pro, use a pair of external
oscillators, to generate 44.1 or 48kHz multiples and are forced to
resort to hacks [1] in order to support 24-bit data without ending up
with fractional dividers.  This patch allows the machine driver to use
32-bit frames for 24-bit data to avoid such issues.

Although the datasheet (p. 15) seems to suggest that only a handful
of ratios are supported, it's not very explicit about it, so we allow
the full range of values supported by the underlying register in the
callback, to avoid needlessly rejecting potentially usable
configurations.

[1] http://mailman.alsa-project.org/pipermail/alsa-devel/2018-December/143442.html

Signed-off-by: Dimitris Papavasiliou <dpapavas@gmail.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ccc8d6c7b6d2f521a4b10c7f6d027f46c7a686bf
 https://github.com/broonie/asoc for-next)

BUG=b:110025789
TEST=Test Audio use cases.

Change-Id: Ie69b4f5d0bdd80788c1af0f0b67e9b1155515c34
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>