CHROMIUM: ASoC:hdac_hda: use correct format to setup hda codec
The current implementation of the hdac_hda codec results in zero-valued
samples on capture and noise with headset playback when SOF is used on
platforms with an on-board HDaudio codec. This is root-caused to SOF
using be_hw_params_fixup, and the prepare() call using invalid runtime
fields to determine the format.

This patch moves the format handling to the hw_params() callback, as
done already for hdac_hdmi, to make sure the fixed-up information is
taken into account but keeps the codec initialization in prepare() as
the stream_tag is only available at that time. Moving everything in the
prepare() callback is possible but the code is less elegant so this
two-step solution was chosen.

The solution was tested with the SST driver with no regressions, and all
the issues with SOF playback and capture are solved.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>
(cherry picked from commit dae8b9d52c599175b3ba3da8fea3a9b67e651dd3
 https://github.com/thesofproject/linux topic/sof-v4.14-rebase)

BUG=b:110025789
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: Idd78b3f931a9fed794290871d66483050a255211
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>