UPSTREAM: ASoC: core: Init pcm runtime work early to avoid warnings
There are cases where we fail before we reach soc_new_pcm which would
init the workqueue. When we fail we attempt to flush the queue which
generates warnings from the workqueue subsystem when we have not inited
the queue. Solution is to use a proxy function to get around this issue.

Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Takashi Iwai <tiwai@suse.de>
Link: https://lore.kernel.org/r/20191203173007.46504-1-cujomalainey@chromium.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 4bf2e385aa59c2fae5f880aa25cfd2b470109093)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:147436144
TEST=Audio playback via Speaker, Headphone and HDMI/DP

Change-Id: I048ac2fe4235b6f515f488a2c851e016b7e9021e