soundwire: intel: Fix a io timeout issue if SDW_INTEL_CLK_STOP_BUS_RESET is set
If two masters are working and one of them become suspended and
activated again, this master will be failed to resume in
SDW_INTEL_CLK_STOP_BUS_RESET mode.

The reason is: on intel platform, there are four masters in the same
power domain. If one master is active, all other masters are always
powered on.  Each master will be suspend to clock stop when it is
inactive. And in resume function each master will be initialized
because it assumes it is powered off. But if the master is not powered
off, it should be in clock stop state, and the normal initialization
will be failed in this state.

Now check clock status and don't initialize master if it is in clock
stop mode.

Signed-off-by: Rander Wang <rander.wang@intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit bbcd9d0abf9c8ab77e6519daa80054609803806b)