CHROMIUM: soundwire: intel: refine runtime pm for SDW_INTEL_CLK_STOP_BUS_RESET
When all the links are suspended, the HDaudio controller may suspend
and the power rails to the SoundWire IP may be disabled, requiring a
complete re-initialization/enumeration on resume. However, if one or
more Masters remained active, the HDaudio controller will remain active
and the power rails will remain enabled. As a result, during the link
resume step we can check if the context was preserved by verifying if
the clock was stopped, and avoid doing a complete bus reset and
re-enumeration.

Signed-off-by: Rander Wang <rander.wang@intel.com>
(cherry picked from commit 24987d3207ae0549b205919e25fac859a9204039
https://github.com/thesofproject/linux.git topic/sof-dev-rebase)

BUG=b:145960362
TEST=Audio Playback via Speaker/Headset using Soundwire Daughter Card
TOREVERT: Once upstream solution lands revert this.

Change-Id: Ifbaf828e3203f39b99ac4e1b607816dca3caedbc