CHROMIUM: soundwire: intel: follow documentation sequences for SHIM registers
Somehow the existing code is not aligned with the steps described in
the documentation, refactor code and make sure the register
programming sequences are correct.

This includes making sure SHIM_SYNC is programmed only once, before
the first master is powered on. We use a 'shim_mask' field, shared
between all links and protected by a mutex, to deal with power-up and
power-down sequences.

Note that the SYNCPRD value is tied only to the XTAL value and not the
current bus frequency or the frame rate.

Signed-off-by: Rander Wang <rander.wang@intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit 0429d418c96d3ae4131117d2fefe82ffa5ee5851
https://github.com/thesofproject/linux.git topic/sof-dev-rebase)

BUG=b:145960362
TEST=Audio Playback via Speaker/Headset using Soundwire Daughter Card
TOREVERT: Once upstream solution lands revert this.

Change-Id: I7dd4ac0b50ad6c1a6740fa1d079fcd00850d1cf8