UPSTREAM: ASoC: core: allow a dt node to provide several components
At the moment, querying the dai_name will stop of the first component
matching the dt node. This does not allow a device (single dt node) to
provide several ASoC components which could then be used through DT.

This change let the search go on if the xlate function of the component
returns an error, giving the possibility to another component to match
and return the dai_name.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Link: https://lore.kernel.org/r/20200213155159.3235792-2-jbrunet@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 1dfa5a5ab34560fd9647083f623d19705be2e706)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:153422295
TEST=No audio regression on TGL/JSL devices

Change-Id: Iaabe1bc9310e1fc08e00587483363dfeed372ebe
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>