CHROMIUM: soundwire: cadence: allocate/free dma_data in set_sdw_stream
The current memory allocation is somewhat strange: the dma_data is
allocated in set_sdw_stream, but released in the intel DAI
shutdown. This no longer works with the multi-cpu implementation,
since the dma_data is released in the dai shutdown which takes place
before the dailink shutdown.

Move to a more symmetric allocation where the dma_data is allocated
with non-NULL SoundWire stream, and conversely released when a NULL
stream is provided - for consistency with the stream startup and
shutdown operations.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>

BUG=b:145960362
TEST=Audio Playback via Speaker/Headset using Soundwire Daughter Card
TOREVERT: Once upstream solution lands revert this.

Change-Id: I0d324466e5f8f4e11faa04e074d3cfdd1aab7086