ASoC: Intel: sdw_711_rt1308_rt715: check endpoint descriptor to get BE DAI number
SDW link addr has been updated to include the endpoint
descriptor which also holds information about whether
the slave should be considered standalone or belonging
to an aggregated group. When there are no aggregated
slave groups, there is a 1:1 mapping between the number
of slaves and the number of BE DAI links. But when slaves
are aggregated in groups, the number of BE DAI links equals
the number of unique groups plus the number of standalone
slaves. Also, when slaves belonging to the same aggregated
group are distributed across multiple SDW links, they require
multiple CPU DAIs depending on the number of links

Tested on Comet Lake H & U, Tiger Lake RVP with 1308 in I2S and sdw
mode.

Signed-off-by: randerwang <rander.wang@linux.intel.com>