UPSTREAM: ASoC: SOF: add core id to sof_ipc_comp
Adds core id to sof_ipc_comp. The intention of this change
is to inform FW on which core that particular component
should run. Right now core id is only passed when pipeline
is created, which is not flexible enough and doesn't allow
for FW to handle this the right way.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Tomasz Lauda <tomasz.lauda@linux.intel.com>
Link: https://lore.kernel.org/r/20200228231850.9226-5-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 1a2289fdf678b780b2d68f408523c09b7074982e)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:153422295
TEST=No audio regression on TGL/JSL devices

Change-Id: Ib02005c200f7cea5110fd77d181bb47917f44c7f
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>