UPSTREAM: ASoC: soc-acpi: expand description of _ADR-based devices
For SoundWire, we need to know if endpoints needs to be 'aggregated'
(MIPI parlance, meaning logically grouped), e.g. when two speaker
amplifiers need to be handled as a single logical output.

We don't necessarily have the information at the firmware (BIOS)
level, so add a notion of endpoints and specify if a device/endpoint
is part of a group, with a position.

This may be expanded in future solutions, for now only provide a group
and position information.

Since we modify the header file, change all existing upstream tables
as well to avoid breaking compilation/bisect.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20200325215027.28716-2-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 004bd4163104e4d8b6c1433b31ead10a69c69845)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>

BUG=b:153422295
TEST=No audio regression on TGL/JSL devices

Change-Id: I6e2417efa90598df1185d94f320c2d4a532fb7b8
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>